+++ R/estimate_contrasts.frequentist.R
@@ -15,7 +15,7 @@
 #' estimate_contrasts(model)
 #' estimate_contrasts(model, fixed = "Petal.Width")
 #' estimate_contrasts(model, modulate = "Petal.Width", length = 4)
-#'
+#' estimate_contrasts(model, levels = "Petal.Width", length = 4)
 #'
 #' if (require("lme4")) {
 #'   data <- iris
@@ -31,12 +31,20 @@
 #' @importFrom bayestestR describe_posterior
 #' @export
 estimate_contrasts.lm <- function(model, levels = NULL, fixed = NULL, modulate = NULL, transform = "none", length = 10, standardize = TRUE, standardize_robust = FALSE, ci = 0.95, adjust = "holm", ...) {
-  estimated <- .emmeans_wrapper(model, levels = levels, fixed = fixed, modulate = modulate, transform = transform, length = length, type = "contrasts")
-  contrasts <- emmeans::contrast(estimated$means, method = "pairwise", adjust = adjust)
+  args <- .guess_arguments(model, levels = levels, fixed = fixed, modulate = modulate)
+
+  estimated <- .emmeans_wrapper(model, levels = args$levels, fixed = args$fixed, modulate = args$modulate, transform = transform, length = length, ...)
+  contrasts <- emmeans::contrast(estimated,
+    by = c(.clean_argument(args$fixed), .clean_argument(args$modulate)),
+    method = "pairwise",
+    adjust = adjust,
+    ...
+  )
+
 
   # Summary
   contrasts <- as.data.frame(merge(as.data.frame(contrasts), stats::confint(contrasts, level = ci, adjust = adjust)))
-  contrasts <- .clean_emmeans_frequentist(contrasts)
+  contrasts <- .clean_names_frequentist(contrasts)
 
   # Reorder columns
   order_SE <- grep("SE", names(contrasts))
@@ -44,7 +52,7 @@ estimate_contrasts.lm <- function(model, levels = NULL, fixed = NULL, modulate =
   contrasts <- cbind(contrasts[c(1:order_SE)], contrasts[col_order[col_order %in% names(contrasts)]])
 
   # Standardized differences
-  if (standardize) {
+  if (standardize & transform != "response") {
     contrasts <- cbind(contrasts, .standardize_contrasts(contrasts, model, robust = standardize_robust))
   }
 
@@ -53,7 +61,8 @@ estimate_contrasts.lm <- function(model, levels = NULL, fixed = NULL, modulate =
   names <- contrasts$contrast
 
   # Separate Contrasts from Others
-  if (!is.null(fixed) | !is.null(modulate)) {
+  # if (!is.null(fixed) | !is.null(modulate)) {
+  if (!is.null(args$modulate)) {
     others <- strsplit(as.character(names), ", ")
     others <- data.frame(do.call(rbind, others))
     names(others) <- unlist(sapply(others, .find_name_level))
@@ -68,11 +77,7 @@ estimate_contrasts.lm <- function(model, levels = NULL, fixed = NULL, modulate =
 
 
   # Format contrasts names
-  levelcols <- strsplit(as.character(levelcols$Contrast), " - ")
-  levelcols <- data.frame(do.call(rbind, levelcols))
-  names(levelcols) <- c("Level1", "Level2")
-  levelcols$Level1 <- gsub(",", " - ", levelcols$Level1)
-  levelcols$Level2 <- gsub(",", " - ", levelcols$Level2)
+  levelcols <- .format_names_contrasts(model, levelcols, transform = transform)
 
   contrasts$contrast <- NULL
   if (nrow(others) != nrow(levelcols)) {
@@ -84,9 +89,9 @@ estimate_contrasts.lm <- function(model, levels = NULL, fixed = NULL, modulate =
   attributes(contrasts) <- c(
     attributes(contrasts),
     list(
-      levels = estimated$levels,
-      fixed = estimated$fixed,
-      modulate = estimated$modulate,
+      levels = args$levels,
+      fixed = args$fixed,
+      modulate = args$modulate,
       transform = transform,
       ci = ci,
       adjust = adjust,
@@ -101,3 +106,6 @@ estimate_contrasts.lm <- function(model, levels = NULL, fixed = NULL, modulate =
 
 #' @export
 estimate_contrasts.merMod <- estimate_contrasts.lm
+
+#' @export
+estimate_contrasts.glmmTMB <- estimate_contrasts.lm