+++ R/estimate_response.bayesian.R
@@ -60,13 +60,13 @@ estimate_response <- function(model, data = NULL, transform = "response", random
 #' }
 #' @return A dataframe of predicted values.
 #' @export
-estimate_response.stanreg <- function(model, data = NULL, transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "response", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.89, ci_method = "hdi", ...) {
+estimate_response.stanreg <- function(model, data = NULL, transform = "response", random = TRUE, length = 25, preserve_range = TRUE, predict = "response", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.95, ci_method = "hdi", ...) {
 
   # Checks
-  if (any(class(model) == "stanreg") & !requireNamespace("rstanarm", quietly = TRUE)) {
+  if (inherits(model, "stanreg") & !requireNamespace("rstanarm", quietly = TRUE)) {
     stop("This function needs `rstanarm` to be installed.")
   }
-  if (any(class(model) == "brmsfit") & !requireNamespace("brms", quietly = TRUE)) {
+  if (inherits(model, "brmsfit") & !requireNamespace("brms", quietly = TRUE)) {
     stop("This function needs `brms` to be installed.")
   }
 
@@ -76,23 +76,30 @@ estimate_response.stanreg <- function(model, data = NULL, transform = "response"
 
   # Predict link or response
   if (predict == "link" && !insight::model_info(model)$is_ordinal) {
-    if (any(class(model) == "brmsfit")) {
-      posteriors <- brms::posterior_linpred(model, newdata = data, re.form = args$re.form, seed = seed, draws = draws, scale = args$transform)
-    } else{
-      posteriors <- rstanarm::posterior_linpred(model, newdata = data, re.form = args$re.form, seed = seed, draws = draws, transform = args$transform)
+    if (isTRUE(args$transform)) {
+      posteriors <- rstanarm::posterior_epred(model, newdata = data, re.form = args$re.form, seed = seed, nsamples = draws, draws = draws)
+    } else {
+      posteriors <- rstanarm::posterior_linpred(model, newdata = data, re.form = args$re.form, seed = seed, nsamples = draws, draws = draws)
     }
   } else {
-    if (any(class(model) == "brmsfit")) {
-      posteriors <- brms::posterior_predict(model, newdata = data, re.form = args$re.form, seed = seed, draws = draws, transform = NULL)
-    } else{
+    if (inherits(model, "brmsfit")) {
+      posteriors <- brms::posterior_predict(model, newdata = data, re.form = args$re.form, seed = seed, nsamples = draws, transform = NULL)
+    } else {
       posteriors <- rstanarm::posterior_predict(model, newdata = data, re.form = args$re.form, seed = seed, draws = draws, transform = "response")
     }
-
   }
 
   # Summary
   prediction <- .summarize_posteriors(as.data.frame(posteriors, stringsAsFactors = FALSE), ci = ci, centrality = centrality, ci_method = ci_method, test = NULL, rope_range = NULL)
   prediction$Parameter <- NULL
+
+  # Rename
+  var <- names(prediction)[grepl(paste0(centrality, collapse = "|"), tolower(names(prediction)))]
+  if (length(var) == 1) {
+    names(prediction)[names(prediction) == var] <- "Predicted"
+  } else{
+    names(prediction)[names(prediction) %in% var] <- paste0("Predicted_", names(prediction)[names(prediction) %in% var])
+  }
 
   # Draws
   if (keep_draws == TRUE | !is.null(draws)) {
@@ -154,14 +161,15 @@ estimate_link <- function(model, data = "grid", transform = "response", random =
 
 #' @rdname estimate_response.stanreg
 #' @export
-estimate_link.stanreg <- function(model, data = "grid", transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "link", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.89, ci_method = "hdi", ...) {
+estimate_link.stanreg <- function(model, data = "grid", transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "link", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.95, ci_method = "hdi", ...) {
   estimate_response(model, data = data, transform = transform, random = random, length = length, preserve_range = preserve_range, predict = predict, keep_draws = keep_draws, draws = draws, seed = seed, centrality = centrality, ci = ci, ci_method = ci_method, ...)
 }
 
 
+
 #' @rdname estimate_response.stanreg
 #' @export
-estimate_response.data.frame <- function(model, data = NULL, transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "link", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.89, ci_method = "hdi", ...) {
+estimate_response.data.frame <- function(model, data = NULL, transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "link", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.95, ci_method = "hdi", ...) {
 
   # Try retrieve model from data
   if ((is.null(data) | all(data == "grid")) & !is.null(attributes(model)$model)) {
@@ -174,7 +182,7 @@ estimate_response.data.frame <- function(model, data = NULL, transform = "respon
 
 #' @rdname estimate_response.stanreg
 #' @export
-estimate_link.data.frame <- function(model, data = "grid", transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "link", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.89, ci_method = "hdi", ...) {
+estimate_link.data.frame <- function(model, data = "grid", transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "link", keep_draws = FALSE, draws = NULL, seed = NULL, centrality = "median", ci = 0.95, ci_method = "hdi", ...) {
 
   # Try retrieve model from data
   if ((is.null(data) | all(data == "grid")) & !is.null(attributes(model)$model)) {
@@ -193,4 +201,3 @@ estimate_response.brmsfit <- estimate_response.stanreg
 
 #' @export
 estimate_link.brmsfit <- estimate_link.stanreg
-