+++ R/estimate_response.frequentist.R
@@ -16,8 +16,13 @@
 #' }
 #' @return A dataframe of predicted values.
 #' @export
-estimate_response.glm <- function(model, data = NULL, transform = "response", random = FALSE, length = 25, preserve_range = TRUE, predict = "response", ci = 0.95, ...) {
+estimate_response.glm <- function(model, data = NULL, transform = "response", random = TRUE, length = 25, preserve_range = TRUE, predict = "response", ci = 0.95, ...) {
   args <- .estimate_response_init(model, data, transform, random, length, preserve_range, predict, ...)
+
+  # Avoid failure in predict()
+  if(is.null(ci)){
+    ci <- 0
+  }
 
   prediction <- predict_wrapper(model,
     newdata = args$data,
@@ -28,11 +33,16 @@ estimate_response.glm <- function(model, data = NULL, transform = "response", ra
     ...
   )
 
-
   # Rename
   names(prediction)[names(prediction) == "fit"] <- "Predicted"
   names(prediction)[names(prediction) == "lwr"] <- "CI_low"
   names(prediction)[names(prediction) == "upr"] <- "CI_high"
+
+  # Restore ci=NULL
+  if(ci == 0){
+    prediction$CI_low <- prediction$CI_high <- ci <- NULL
+  }
+
 
 
   # Add predictors
@@ -76,6 +86,11 @@ estimate_link.lm <- estimate_link.glm
 estimate_response.merMod <- estimate_response.glm
 #' @export
 estimate_link.merMod <- estimate_link.glm
+
+#' @export
+estimate_response.glmmTMB <- estimate_response.glm
+#' @export
+estimate_link.glmmTMB <- estimate_link.glm
 
 #' @export
 estimate_response.polr <- estimate_response.glm