+++ R/utils.R
@@ -23,7 +23,6 @@
 #' @examples
 #' as.numeric_ifnumeric(c("1", "2"))
 #' as.numeric_ifnumeric(c("1", "2", "A"))
-#'
 #' @return Numeric
 #' @export
 as.numeric_ifnumeric <- function(x) {
@@ -39,8 +38,8 @@ as.numeric_ifnumeric <- function(x) {
 
 
 #' @keywords internal
-.remove_name_level <- function(x) {
-  name <- .find_name_level(x)
+.remove_name_level <- function(x, ...) {
+  name <- .find_name_level(x, ...)
   x <- sub(name, "", x)
   x <- trimws(x)
   x
@@ -53,7 +52,11 @@ as.numeric_ifnumeric <- function(x) {
 
 
 #' @keywords internal
-.find_name_level <- function(x, data, fixed, modulate) {
+.find_name_level <- function(x) {
+  if (length(unique(x)) == 1) {
+    return("Contrast")
+  }
+
   splitted <- strsplit(as.character(x), " ")
   splitted <- data.frame(do.call(rbind, splitted), stringsAsFactors = FALSE)
   uniques <- sapply(splitted, unique)